osd/scrub: implement urgency_t::repairing
to be used for repair scrubs triggered when a shallow
scrub ends with repairable errors.

The deep scrub that follows should not have the same
urgency as an operator-requested repair scrub. It should
not, for example, be allowed to run when the OSD is
performing a recovery, osd_scrub_during_recovery is false,
and osd_repair_during_recovery is set.

Signed-off-by: Ronen Friedman <rfriedma@redhat.com>